+++ src/hover.ts
@@ -4,10 +4,7 @@ import logger from "./logger";
 import { getMathSignatures, getMathDistributions } from "./callStanc";
 import { getDocumentationForFunction } from "./documentation";
 
-const functionSignatureMap: Map<
-  string,
-  Array<string | vscode.MarkdownString>
-> = new Map();
+const functionSignatureMap: Map<string, vscode.MarkdownString> = new Map();
 const distributionToFunctionMap: Map<string, string> = new Map();
 
 function setupSignatureMap() {
@@ -16,15 +13,24 @@ function setupSignatureMap() {
   for (const line of lines) {
     const [name] = line.split("(", 1);
     if (functionSignatureMap.has(name)) {
-      functionSignatureMap.get(name)?.push(`\`${line}\``);
+      functionSignatureMap.get(name)?.appendCodeblock(line, "stan");
     } else {
-      functionSignatureMap.set(name, [
-        getDocumentationForFunction(name),
-        "Available signatures",
-        `\`${line}\``,
-      ]);
+      functionSignatureMap.set(
+        name,
+        getDocumentationForFunction(name)
+          .appendMarkdown("\n\n**Available signatures**:")
+          .appendCodeblock(line, "stan"),
+      );
     }
   }
+  // not technically 'functions', but still useful
+  functionSignatureMap.set("print", getDocumentationForFunction("print"));
+  functionSignatureMap.set("reject", getDocumentationForFunction("reject"));
+  functionSignatureMap.set(
+    "fatal_error",
+    getDocumentationForFunction("fatal_error"),
+  );
+  functionSignatureMap.set("target", getDocumentationForFunction("target"));
 
   const mathDistributions = getMathDistributions();
   const distLines = mathDistributions.split("\n");
@@ -43,20 +49,19 @@ function provideHover(
   document: vscode.TextDocument,
   position: vscode.Position,
 ): vscode.Hover | undefined {
-  const fnRange = document.getWordRangeAtPosition(position, /[\w_]+\s*\(/);
-  if (!fnRange) {
-    return undefined;
+  let fnName: string | undefined = undefined;
+  const distRange = document.getWordRangeAtPosition(position, /~\s*[\w_]+/);
+  if (distRange) {
+    const distName = document.getText(distRange).slice(1).trim();
+    fnName = getDistributionName(distName);
   }
-  let fnName = document.getText(fnRange).slice(0, -1).trim();
-  const distName = getDistributionName(fnName);
-  if (distName) {
-    fnName = distName;
+  if (!fnName) {
+    const fnRange = document.getWordRangeAtPosition(position, /[\w_]+\s*\(/);
+    if (!fnRange) return undefined;
+    fnName = document.getText(fnRange).slice(0, -1).trim();
   }
-
-  const signatures = functionSignatureMap.get(fnName);
-  if (!signatures) {
-    return undefined;
-  }
+  let signatures = functionSignatureMap.get(fnName);
+  if (!signatures) return undefined;
   return new vscode.Hover(signatures);
 }